refactor: Split transactionService into multiple files.
