+++ frontend/src/components/accounts.js
@@ -1,6 +1,6 @@
 import { React, useState, useEffect } from "react";
 import { Formik, Form, Field } from "formik";
-import transactionService from "../services/transactionService";
+import transactionService from "../services/transactionService/transactionService";
 import NoDataCard from "./core/nodata";
 import "./accounts.scss";
