+++ frontend/src/components/dashboard/expenses.js
@@ -1,6 +1,6 @@
 import { Formik, Form, Field } from "formik";
 import React, { useEffect, useState } from "react";
-import transactionService from "../../services/transactionService";
+import transactionService from "../../services/transactionService/transactionService";
 import DatePicker from "react-datepicker";
 import "react-datepicker/dist/react-datepicker.css";
 import "./expenses.scss";
