+++ frontend/src/components/dashboard/incomes.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import transactionService from "../../services/transactionService";
+import transactionService from "../../services/transactionService/transactionService";
 import DatePicker from "react-datepicker";
 import "react-datepicker/dist/react-datepicker.css";
 import { Formik, Form, Field } from "formik";
