+++ frontend/src/components/templates/templateGroupForm.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import transactionService from "../../services/transactionService";
+import transactionService from "../../services/transactionService/transactionService";
 import { Formik, Form, Field } from "formik";
 import "./templates.scss";
