+++ frontend/src/services/transactionService/transferService.js
@@ -1,0 +1,41 @@
+import axios from "axios";
+
+const BASE_URL = "http://localhost:8000/transactions";
+
+async function getAllUserTransfers() {
+    const token = JSON.parse(localStorage.getItem("authToken"));
+    const config = {
+        headers: {
+            Authorization: token,
+        },
+    };
+
+    const response = await axios.get(`${BASE_URL}/alltransfers`, config);
+
+    if (response.status === 200) {
+        return response.data;
+    } else {
+        return "no data";
+    }
+}
+
+async function addTransfer(payload) {
+    const token = JSON.parse(localStorage.getItem("authToken"));
+    const config = {
+        headers: {
+            Authorization: token,
+        },
+    };
+
+    const response = await axios.post(`${BASE_URL}/add`, payload, config);
+    if (response.status === 201) {
+        return response.data;
+    } else {
+        alert("Error adding transfer.");
+    }
+}
+
+export default {
+    getAllUserTransfers,
+    addTransfer,
+};
